Update HW4
I got a little lost among the commands, I couldn't recognize everything you are doing here. For example, it took me a while to understand what is pargrai, pargrau, GM, ri, ru and what is happening with them, I wasn't sure about the results for the parametric intervals on 3.1 either. I think it would be helpful if you set the step-by-step process you are taking or making the functions more clear.
Instead of plugging in numbers, your code could be more functional for other studies if you plugging variables. For example, instead of using "lowgr1<- ogr[5]+0.025*(ogr[6]-ogr[5])", you could use "lowgr1<- exp(mean.ln.grano-t.value*sqrt(var.grano/n))", in which you can plug in the numbers of the variables, for the lower or upper interval, or use them in other functions, without having to change them for each function all the time, so you can just change the variables and the rest is already done.